split HierarchyView.tsx
